fix: controller ve servis sınıfları düzenlendi
