+++ index.php
@@ -5,8 +5,8 @@ require 'core/model/route.php';
 require 'controller/fibonacci-number-controller.php';
 require 'controller/prime-number-controller.php';
 require 'controller/home-controller.php';
-require 'model/fibonacci-number-calculator.php';
-require 'model/prime-number-calculator.php';
+require 'services/fibonacci-number-calculator.php';
+require 'services/prime-number-calculator.php';
 require 'core/request-parser.php';
 require 'config/database.php';
 require 'controller/user-controller.php';
