+++ services/fibonacci-number-calculator.php
@@ -1,5 +1,5 @@
 <?php
-require 'model/abstract-calculator.php';
+require 'services/abstract-calculator.php';
 class FibonacciNumberCalculator extends AbstractCalculator
 {
     private $step;
