+++ services/product-service.php
@@ -2,20 +2,35 @@
 class ProductService
 {
     /**
-     * @var Product[] $productObject
+     * @var Product[] $productList
      * 
      */
 
-    private array $productObject;
+    private array $productList;
+    private Database $database;
 
     public function __construct()
     {
-        $productController = new ProductController();
-        $this->productObject = $productController->getAllProduct();
+        $this->database = new Database();
+        $this->productList = [];
     }
 
     public function getProductList(): array
     {
-        return $this->productObject;
+        try {
+            $connection = $this->database->connection;
+            $query = "SELECT * FROM product";
+            $products = mysqli_query($connection, $query);
+            if ($products->num_rows > 0) {
+                while ($row = $products->fetch_assoc()) {
+
+                    $this->productList[] = new Product($row["id"], $row["name"], $row["price"]);
+                }
+                mysqli_close($connection);
+                return $this->productList;
+            }
+        } catch (Exception $e) {
+            echo $e->getMessage();
+        }
     }
 }
